+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file. Be aware th
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased] 2022-09-21
+### Changed
+- Changed the Wrath nameplate aura filter to include actual auras.
+
 ## [2.0.47-RC] 2022-09-21
 ### Added
 - Added Druid/Rogue/Vehicle Combo Points, Mage Arcane Charges, Monk Chi, Paladin Holy Power and Warlock Soul Shards.